Renamed python files to be in line with best python practices.  Moved tests into aida_interchange to deal with import issues. Updated Readme and added a requirements.txt file for dependencies of python virtual environment.
