+++ python/aida_interchange/aifutils.py
@@ -4,7 +4,7 @@ from abc import ABCMeta, abstractmethod
 
 from rdflib import URIRef, RDF, Graph, BNode, Literal, XSD
 
-from aida_interchange.aida_rdf_ontologies import AIDA_ANNOTATION
+from aida_rdf_ontologies import AIDA_ANNOTATION
 from rdflib.plugins.sparql import prepareQuery
 
 """
