+++ python/aida_interchange/ldc_time_component.py
@@ -1,7 +1,7 @@
 from enum import Enum
 from rdflib import Graph, Literal, XSD, URIRef
-from aida_interchange.aida_rdf_ontologies import AIDA_ANNOTATION
-from aida_interchange import aifutils
+from aida_rdf_ontologies import AIDA_ANNOTATION
+import aifutils
 
 class LDCTimeType(Enum):
     ON = 1
