+++ python/aida_interchange/tests/invalid_examples.py
@@ -3,8 +3,8 @@ import sys
 sys.path.append('../')
 from io import BytesIO
 from rdflib import URIRef, Literal, XSD, BNode, RDF
-from aida_interchange.aida_rdf_ontologies import AIDA_PROGRAM_ONTOLOGY, AIDA_ANNOTATION
-from aida_interchange import aifutils
+from aida_rdf_ontologies import AIDA_PROGRAM_ONTOLOGY, AIDA_ANNOTATION
+import aifutils
 
 class InvalidExamples(unittest.TestCase):
