+++ python/aida_interchange/tests/scaling_test.py
@@ -1,7 +1,7 @@
 import sys
 sys.path.append('../')
-from aida_interchange.aida_rdf_ontologies import SEEDLING_TYPES_NIST
-from aida_interchange import aifutils
+from aida_rdf_ontologies import SEEDLING_TYPES_NIST
+import aifutils
 import random
 from rdflib import URIRef
 import os.path
