+++ python/requirements.txt
@@ -1,0 +1,1 @@
+rdflib==4.2.2
